Done some minute changes in all frontend files

+++ public/src/components/Logout.jsx
@@ -1,25 +1,29 @@
 import React from "react";
-import styled from "styled-components";
 import { useNavigate } from "react-router-dom";
-import { Axios } from "axios";
 import { BiPowerOff } from "react-icons/bi";
+import styled from "styled-components";
+import axios from "axios";
+import { logoutRoute } from "../utils/APIRoutes";
 
-const Logout = () => {
+export default function Logout() {
   const navigate = useNavigate();
-
   const handleClick = async () => {
-    localStorage.clear();
-    navigate("/login");
+    const id = await JSON.parse(
+      localStorage.getItem(process.env.REACT_APP_LOCALHOST_KEY)
+    )._id;
+    const data = await axios.get(`${logoutRoute}/${id}`);
+    if (data.status === 200) {
+      localStorage.clear();
+      navigate("/login");
+    }
   };
-
   return (
     <Button onClick={handleClick}>
       <BiPowerOff />
     </Button>
   );
-};
+}
 
-export default Logout;
 const Button = styled.button`
   display: flex;
   justify-content: center;
